Automatically filter out newline when pasting from Clipboard. fixes #97
Newlines (both regular and windows ones) are now automatically
removed when Clipboard content is being pasted into a Textfield. This
should fix the issue.

I agree that the other part of the issue (fixing the handling of
multi-line selections) needs to be fixed as well. But this is more
involved and should be worked on separately.