workaround for issue #73 added
The KeyboardInputEvent constructor sets the character instance variable
to 0 when the control-key is pressed. This prevents characters you can
only reach with the help of control and alt, like AltGr+Q on Windows to
insert a @ character for certain keyboard layouts.

To prevent this issue I've now removed the "if (controlDown)" check in
the constructor. This means any character code pressed while any of the
special keys is down (alt, shift, control) will be processed as
a regular key.

However, the InputMapping implementation, f.i. TextfieldInputMapping
will handle some control+key combinations like ctrl+A or ctrl+C and so
correctly and will prevent any character input into the textfield in
that case. Other, unhandled combinations will still find it's way to
the textfield which is not quite right but shouldn't do any harm. For
instance you can now press ctrl-P and a you would see "P" in the
textfield.