method markBatchRenderImageAsUnloaded() added to NiftyImage
This method allows you to modify a backing texture externally and then
notify Nifty that you've changed the texture. Nifty will then upload the
image to the texture atlas the next time the image is accessed. Without
this method Nifty would never know about the change of the texture and
would never update the texture atlas with the updated texture. This
method is specific to the new BatchRenderer.