fixes #109: Assigning an id from a custom attribute to a control
resolveParameters() needs to happen before makeFlatControls() since
makeFlatControls() resolves the ids and propagates the id to child
elements. which is ok ... unless someone comes and changes the actual
ids with some $parameter! In that case makeFlatControls() did nothing
since there was no actual id available! This commit seems to fix that.
I'm not sure if something else breaks now although all the standard
demos/examples I've checked seem to work.

This part of Nifty is kinda scary! Note to self: make a better
implementation of all of this for Nifty 2.0!