#136 Unsubscribe all elements in element.onEndScreen()
Up until now only controllers have been automatically unsubscribed when
the element it was attached to ended. This worked for most cases but it
didn't work when you've manually subscribed to events of controls inside
other controls. While Nifty iterated child elements it only
automatically unsubscribed controls for the parent element in that case.

Now all elements in the hierachy of the removed/ended element will be
automatically unsubscribed.