#139 fix display of initial CheckBox state
The CheckBox didn't display its correct state when it was shown after
being hidden.

Example: CheckBox on a hidden Tab and change it's state by code. When
you then display the tab the Checkbox won't show it's correct state.

I've modified the TabsControlDialog in the standard controls demo to
test this.