Minor cleanups.
1) Remove useless, incorrect, outdated comment from Nifty constructor.

2) Remove strange, invalid @created javadoc from
SpecialValuesReplaceTest. It creates a warning when running mvn site on
nifty-core.

3) Rename android module in nifty-examples-libgdx from 'Main Android' to
'NiftyExamples for LibGDX - Android' so that it's clear when building
Nifty with the maven reactor that the module is a submodule of
nifty-examples-libgdx.

4) Rename core submodule in nifty-examples-libgdx from 'Main Core' to
'NiftyExamples for LibGDX - Main Core' so that it's clear when building
Nifty with the maven reactor that the module is a submodule of
nifty-examples-libgdx.

5) Rename desktop submodule in nifty-examples-libgdx from 'Main Desktop'
to 'NiftyExamples for LibGDX - Desktop' so that it's clear when building
Nifty with the maven reactor that the module is a submodule of
nifty-examples-libgdx.

6) Create a logging.properties file for nifty-examples-libgdx, and
configure all logging to Level.WARNING. Remove the static Logger
instance from LibgdxExampleApplication and use a LogManager to load the
logging.properties file instead.

7) Fix an indentation error in nifty-examples-libgdx parent pom file.

8) Fix code style violation in GdxInputSystem by removing unnecessary
uses of the 'this' keyword.