#138: [1.4] childClip inside childClip does not work correctly
Originally if you had a childClip="true" element in another
childClip="true" element the clipping rectangle of the child was ignored
and the first encountered clipping rectangle (childClip="true" element)
was used.

This has now been changed and the following cases are handled:

1. If the clipping rectangle of the child element is completely outside
of its parent clipping rectangle then the clipping rectangle is not
changed.

2. Is the child clipping rectangle completely inside its parent clipping
rectangle then the child clipping rectangle will be used as the new
clipping rectangle.

3. If both rectangles intersect each other the interesection rectangle
will be calculated and used.