Remove internal use of deprecated Element methods.
Remove all uses of all deprecated methods in Element class. The
deprecated methods are still there, but they are not used anymore
internally.

There is also one deprecated method in Element class that is supposed to
delegate to the non-deprecated version, but instead it is duplicating
the non-deprecated version's code. Make this deprecated method delegate
to the non-deprecated version.

Also remove a duplicated word from a javadoc comment in the
getValueAsInt method of SizeValue class.

There are some faux changes present where the entire file appears to
have changed, but it's just due to the removal of  carriage return
characters in the line endings.