Handle children storage corruption in element
The children storage of the element got corrupted in case the ID of a
element got changed after it was added to its parent. In this case it
was likely to happen that the element couldn't be located anymore in
the tree set used to maintain the render order.

Now the elements reports any change of the ID to its parent to give
it the chance to update the rendering order as needed.