Bugfix batched renderer (changes to texture atlas need bind() call!)
Usually when Nifty renders using the batched renderer the texture atlas
is the only texture bound to GL_TEXTURE_2D for the whole time.

However since we allow dynamically creating, modifying the texture atlas
it is possible that calls to: clearAtlasTexture(), addImageToTexture()
and  removeFromTexture() are executed while the texture atlas texture is
not currently bound.

This patch fixes this. Please note that we call bind() directly in the
method and therefore potentially too often. This is not a problem for
clearAtlasTexture() and removeFromTexture() because they will be called
just a few times (or only once for clearAtlasTexture()). For
addImageToTexture() we bind the texture atlas for each call of that
method with is not optimal but shouldn't matter too much (mot verified
in detail tho).

Kudos to Andreas Drewke for figuring this one out!