#71 Optimize sending events when changing constraints
Changing constraints will not immediately publish an event on the
eventbus anymore (with some exceptions noted below). Instead the element
will just be marked and the event will be send on the next layout pass.

Doing it this way is more correct because the event will now be sent
directly after the change has been actually applied (and not at some -
possible long - time before it actually happened).

And when changing multiple constraints you'll now only get a single
Element changed event and not multiple.

There are some changes to an Element that will have a direct effect on
it (and not at the next layout pass) that will be published immediately:

- setParent() when the parent clip area is changed
- setStyle()
- removeStyle()