new moveToRelative() method added to fix a bug in AllExamples example
This was actually an interesting bug to fix because there were several
things interacting in weird ways.

The problem: In the AllExamples menu the selected menu item was animated
and rendered with the TextSize effect making it larger than regular menu
items. However with the latest Nifty 1.4 the selected menu-item was
stuck at it's initial position not being moved around with the rest of
the menu. WTF?

Issues:

First, a couple of weeks back we've cleaned up rendering the way Nifty
keeps track of render states while rendering. There was a lot of
(mostly) unnecessary render state saves happening while rendering which
have been removed. I think this affected the moveTo() call on the
NiftyRenderEngine which sets a current offset being applied when
rendering. The TextSize effect uses moveTo() to adjust the text position
according to the textScale. However, it did that absolutely without
taking into account that the current moveTo could already been modified
by some effect before (like the moving in of the whole menu). I think in
the past this moveTo() state was saved correctly but because of the
change it is not saved correctly.

The second issue was the introduction of the absolute clipping. The menu
was using childClip which will only actually clip if there was no
clipping set before. With absolute clipping enabled and the menu
starting off-screen the clipping rectangle was not changed end the
absolute clipping (whole screen) was still enabled. Because of the move
effect not moving the selected menu item and it was not being clipped it
stayed at its original position looking odd :) Without absolute clipping
the moveTo bug was not really visible because it was clipped.

To fix all of this we've now introduced a new moveToRelative() method
which is used by the TextSize effect now. This will add the moveTo
effect to the one that might already be applied. So the TextSize
position adjusting (because of the new textsize) AND the move effect of
the whole menu will now be added together and applied both.

This is a little hack but it will solve the issue for now.

Future versions of Nifty will have a proper transformation matrix - I
promise ;)