#84 More flexible solution in using control definition parameters
control parameters are now detected using two different ways:

1. the old syntax: $parametername
2. the new syntax: $(parametername)

Please note that the new syntax detects parameters in substrings as
well. For example using: text="Hello $(myText)" will create a new
parameter "myText" and will replace the value given for "myText" in the
substring when the control is being resolved. See the added issue84.xml
in this commit for a working example.