#55 prevent double subscription of annotation-based events
The used EventBus already checked for duplicate subscribers. However,
for the Nifty annotations a ProxySubscriber was used and from what I can
tell there is a little bug in the EventBus implementation for that.

When the EventBus checked for duplicate ProxySubscribers it compared the
new Subscriber with the existing once. That is correct. However, for the
subscribers already in the list it ALWAYS used the "proxiedSubscriber"
but for the new subscriber it used the "proxiedSubscriber" only when it
used "ReferenceStrength.WEAK" othwise it would use the subscriber as-is.
This made it never detect existing subscribers and therefore subscribed
them multiple times (even tho they would reference the same proxied
subscriber in the end).

I think this is not correct. I've changed the ReferenceStrength to WEAK
now which "should" me okay and will fix this issue. From what I've seen
in the EventBus source the ReferenceStrength is only used for that check
and does not have any other side-effects. So this should be okay and
fixes this issue.

Additionally I've added an example for the issue reproducing the issue.