#55 Double subscription of annotation-based events really fixed now
After modifying the EventBus project event processing in Nifty didn't
work correctly anymore. The problem was that the ProxySubscriber class
"NiftyEventAnnotationProcessor.Subscriber" class proxied to the same
object but different methods. With the EventBus-project patch in place
these were detected as duplicates and ignored :)

However, when thinking about it again using the ProxySubscriber is not
really necessary for the NiftyEventAnnotationProcessor, so I've removed
it and now the "NiftyEventAnnotationProcessor.Subscriber" is subscribed
directly. For that to work correctly the Subscriber now has equals() and
hashCode() implementations.

Issue55Main testcase has been added again too.