#223 apply absolute clip only when it is enabled + disableAbsoluteClip()
Until now absolute clipping was enabled by default and was being set to
the whole screen. However this lead to problems when panels with
childClip="true" have been moved outside of the screen. For example the
ControlsDemo has a ScrollPanel demo which animates out of the screen.
With absolute clipping enabled (default) the clipping rectangle of the
ScrollPanel was ignored due to the absolute clipping rectangle was
already enabled. This lead to rendering artifacts. In case of the
ControlDemo the big picture that was used as the area to be scrolled
around was displayed completely without being clipped at all.

This change will now apply the absolute clipping ONLY when it is
actually used, e.g. if someone calls nifty.setAbsoluteClip(). If this
method is not called then from now on no absolute clipping is being
applied anymore which fixes the rendering problems in the ControlsDemo.

I'm aware that there might be other problems when absolute clipping is
enabled but I can't fix them right now.

Please Note: Once you enable absolute clipping by a call to
nifty.setAbsoluteClip() it will stay enabled unless you disable it again
with a call to the new method disableAbsoluteClip().