#100 basic support added for changing locale at runtime (text only)
This commit adds support for changing text (TextRenderer) on the fly
when changing the locale using nifty.setLocale(). An example to see that
working has been added as Issue100Main.

The "trick" to make this working is to store the unmodified original
text value inside of the TextRenderer and use it when the locale has
been changed to reinit the TextRenderer with the original value applied.
The TextRenderer already stored the original text value and it already
translated it using the SpecialValuesReplace code! However, so far it
was always called with the already localized text.

To change that we now store in the attributes map additionally to the
translated value the original value. This way when the TextRenderer
is being initialized we can use the original untranslated value!

When the locale is changed using the method nifty.setLocale() a
NiftyLocaleChangedEvent is being published. The TextRenderer now
subscribes to that event and will simply reinitialize itself using
the original text value stored (which will automatically will be
localized using a call to SpecialValuesReplace which was already
in place).

A similar approach could be used for other properties but for the
time being this is only applied to the text attribute of a TextRenderer.