#116 niftyMethodInvokerDebugEnabled added to not supress exceptions
You can now control if Exceptions that occur when the NiftyMethodInvoker
catches an Exception while executing any of the standard interact
handlers are being only logged (the default behaviour) or will be
rethrown so that the whole application crashes (set
niftyMethodInvokerDebugEnabled to true for this mode).

Additionally the logging of the exceptions have been improved, using the
logger framework to log the exception.