Fix NiftyStandardInputEvent.Activate bug.
NiftyStandardInputEvent.Activate only simulates a mouse click, but not a
mouse release. This causes a bug where only listening for
ButtonReleasedEvent's but not ButtonClickedEvent's for a specific button
will cause input mappings to NiftyStandardInputEvent.Activate to not
work with that button. This results in undesirable & unexpected behavior
such as pressing the enter key on a currently focused button does not do
anything.

The solution is to make NiftyStandardInputEvent.Activate simulate a
mouse click and a mouse release, so that when
NiftyStandardInputEvent.Activate is triggered, a ButtonReleasedEvent is
fired in addition to the ButtonClickedEvent.